refactor: split YahooClient across per-concern mixin files
The previous client.py was 804 lines and rolled every concern (quotes,
history, fundamentals, validation, holdings, search, earnings, intraday)
plus all parsing helpers into one file. That's a lot to take in for what
is conceptually one component.

Split into focused files (largest is _parsers.py at 225 lines, client.py
itself is now 134 lines):

- _parsers.py — pure transformations + constants (PERIOD_MAP,
  FUNDAMENTAL_FIELDS, SECTOR_NAMES). No Yahoo I/O, easy to unit-test.
- _base.py — abstract _YahooBase declaring the throttle, caches and
  orchestrator hooks every mixin uses.
- _quotes.py, _history.py, _fundamentals.py, _validation.py,
  _holdings.py, _search.py, _earnings.py, _intraday.py — one mixin per
  concern, 17–106 lines each.
- client.py — YahooClient inherits all mixins, implements __init__,
  _call (orchestrator), _ticker (single Ticker instantiation point), and
  the yahoo_client singleton.

Architectural property preserved: every Yahoo HTTP call still goes
through one yahoo_client with a shared throttle/breaker/caches. Tests
patch ``app.services.yahoo.client.Ticker`` exactly once thanks to the
_ticker indirection — no per-mixin patching needed.

Side benefits while restructuring:
- Collapsed each public method's async/sync pair into a single async
  method that wraps the fetch closure with asyncio.to_thread. The
  internal _quotes_sync/_fundamentals_sync/_earnings_sync/_intraday_sync
  helpers are gone.
- Tests that called those private sync methods now `await` the public
  async methods. test_yahoo_search.py retargets the patch from
  client._yq_search to _search._yq_search; test_intraday.py likewise
  for resolve_currency.

611 tests passing (unchanged).
