+++ backend/tests/services/test_yahoo_quotes.py
@@ -5,7 +5,10 @@ from unittest.mock import MagicMock, patch
 import pytest
 
 from app.services.yahoo import yahoo_client
-from app.services.yahoo.client import _parse_quotes, _sanitize_float
+from app.services.yahoo._parsers import (
+    parse_quotes as _parse_quotes,
+    sanitize_float as _sanitize_float,
+)
 from app.services.yahoo.rate_limit import crumb_rejected
